CORE-7620 Fix metadata-client.core/copy-metadata-avus
Added missing `user` parameter and removed the `force` parameter from
metadata-client.core/copy-metadata-avus, since it's not currently
supported in the metadata service AVU copy endpoint.
Updated the terrain metadata copy endpoint to not send the `force`
parameter to the data-info endpoint if it did not get a value for it
from the client.