Check for tagged apps when renaming a community.
Updated `PATCH /(admin/)communities/{name}` to check for apps tagged in
that community when the admin is renaming that community.
Added `retag-apps` and `force-rename` params to these endpoints,
and if these params are omitted (or set to `false`) when renaming the
community, an `ERR_EXISTS` will be thrown if any apps are tagged.
Otherwise, any app AVU tags will be updated with the new community name
if the `retag-apps` param is set to `true`.

Updated some `terrain.clients.metadata.raw` functions with
`org.cyverse/metadata-client` @ v3.0.1.
Added `terrain.clients.metadata.raw/find-avus`.
Added a `terrain.communities.metadata.attr` config.