CORE-1554 Fixes for permanent-id-request endpoint emails
Refactored staging and publish data-move functions to pass
`current-user` as a param to lower functions for error emails,
since the binding is not available in async threads.

Renamed the email template keys `username` to `user`,
since that's what the de-mailer service is now using in its templates,
but it also makes more sense since the value is the user's full name,
and not their account username.